fix: get_program_price should now take priority on cost field if it exist

+++ src/aleph/sdk/query/responses.py
@@ -79,7 +79,7 @@ class MessagesResponse(PaginationResponse):
 class PriceResponse(BaseModel):
     """Response from an aleph.im node API on the path /api/v0/price/{item_hash}"""
 
-    required_tokens: float
+    required_tokens: Decimal
     payment_type: str
